Abstracted icns utils
